Delay compression until Inject is requested
Compression is only performed when handling the Inject event. If done this way, it greatly simplifies a lot of what was being done to juggle the compression status.

It does also mean that the size of the compressed file won't be available until it's been injected. But that's a small crime. It's not really a relevant number in most use cases, because it's simply being directly inserted and utilized.