Added an Auto option to SLLZ Version choice
The Auto option is enabled only when the Compression Option dropdown is set to Mirror PAR File.

Usually, Mirror PAR File matches the compression status of the injected file with its match in the original PAR file. With Auto selected, injections match the compression version of the matched file as well.

I think there's a good argument to be made for this behaviour to be the default behaviour, but I'll keep that to a separate commit.