Odds and ends
- Removed the unused function getVersionOfCompressedFiles. It was unnecessary, and operated under the assumption that files were always pre-compressed.


- Changed padding logic in UpdateFileStarts for the corner case in which the next file's length is exactly equal to the padding. The prior version would slot that file into the available space, whereas this version applies the padding and sets the file start to after the padding.

My understanding had been that PS3-era files exhibited that prior behaviour, but on re-inspection I couldn't find files in which that was true, and did find files where the opposite was true (the file was padded to a new block despite the available space being exactly the length of the file).

Importantly, both methods do seem to work without fault, and it would be nice to provide an option in the future. But the default behaviour should always be that loading a PAR and immediately saving a duplicate produces an identical file.